Clean up remaining discrepancies
