+++ Cargo.toml
@@ -25,7 +25,6 @@ ethabi = "18.0.0"
 serde = "1.0"
 hex = "0.4"
 chrono = "0.4"
-serde_json = "1.0"
 rust-kzg-bn254-prover = { git = "https://github.com/Layr-Labs/rust-kzg-bn254.git", branch = "master"}
 ark-serialize = "0.5"
 tonic = { version = "0.12.1", features = ["tls", "channel", "tls-roots"]}
@@ -37,7 +36,7 @@ prost = "0.13.1"
 common = {path = "common"}
 secrecy = "0.8.0"
 alloy = "0.11"
-
+ark-ff = "0.5"
 
 # Always optimize; building and running the guest takes much longer without optimization.
 [profile.dev]
