+++ common/Cargo.toml
@@ -5,3 +5,5 @@ edition = "2021"
 
 [dependencies]
 serde = { workspace = true }
+ark-bn254 = { workspace = true }
+ark-ff = { workspace = true }
