+++ common/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod blob_info;
+pub mod serializable_g1;
