+++ host/src/prove_risc0.rs
@@ -5,7 +5,6 @@ use alloy::{
     signers::local::PrivateKeySigner,
     sol,
 };
-use blob_verification_methods::BLOB_VERIFICATION_GUEST_ELF;
 use risc0_zkvm::ProveInfo;
 use risc0_zkvm::{compute_image_id, sha::Digestible};
 use secrecy::{ExposeSecret, Secret};
